Update loopback.js
On a mac loopbackAnswer contains \\r\\n characters instead of proper line breaks. Because of that initial regex /a=crypto:[1-9]+ .*/g matches everything starting from a=crypto:1 all the way to the end of loopbackAnswer, and when replaced for an empty string it corrupts the json. 

As a solution, I propose to first search-and-replace "mac-specific" crypto patterns (crypto all the way to first \\r\\n sequence) and only after that search-and-replace original patterns (crypto all the way to the end of the line).